Move confinement time enums to data structure
